+++ docs/_ru-sidebar-tree.js
@@ -1003,6 +1003,10 @@ module.exports = {
           path: ridePath + 'advanced/continuation',
         },
         {
+          title: 'Подключение библиотек',
+          path: ridePath + 'advanced/import',
+        },
+        {
           title: 'Нетерпеливая переменная',
           path: rideV5Path + 'variables',
         },